modules/requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
gensim==4.3.3
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.43.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
Pygments==2.18.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
smart-open==7.0.5
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
update-checker==0.18.0
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.16.0
zipfile36==0.1.3
